[executorch][serialization] Rename scalar_type.fbs --> common.fbs
Allow us to add other shared flatbuffer types into `common.fbs`.

- Rename scalar_type.fbs --> common.fbs
- Rename scalar_type.py --> common_schema.py (differentiate from exir/common.py)

Differential Revision: [D65442256](https://our.internmc.facebook.com/intern/diff/D65442256/)

[ghstack-poisoned]
