+++ devtools/bundled_program/schema/bundled_program_schema.fbs
@@ -4,7 +4,7 @@
 // See README.md before modifying this file.
 //
 
-include "scalar_type.fbs";
+include "common.fbs";
 
 namespace bundled_program_flatbuffer;
