+++ devtools/etdump/etdump_schema_flatcc.fbs
@@ -1,4 +1,4 @@
-include "scalar_type.fbs";
+include "common.fbs";
 
 namespace etdump;
