+++ devtools/etdump/targets.bzl
@@ -1,7 +1,7 @@
 load("@fbsource//xplat/executorch/build:runtime_wrapper.bzl", "runtime")
 
-SCALAR_TYPE_STEM = "scalar_type"
-SCALAR_TYPE = SCALAR_TYPE_STEM + ".fbs"
+COMMON_STEM = "common"
+COMMON = COMMON_STEM + ".fbs"
 
 # flatcc
 ETDUMP_STEM_FLATCC = "etdump_schema_flatcc"
@@ -12,9 +12,9 @@ ETDUMP_SCHEMA_FLATCC_BUILDER = ETDUMP_STEM_FLATCC + "_builder.h"
 ETDUMP_SCHEMA_FLATCC_READER = ETDUMP_STEM_FLATCC + "_reader.h"
 ETDUMP_SCHEMA_FLATCC_VERIFIER = ETDUMP_STEM_FLATCC + "_verifier.h"
 
-SCALAR_TYPE_BUILDER = SCALAR_TYPE_STEM + "_builder.h"
-SCALAR_TYPE_READER = SCALAR_TYPE_STEM + "_reader.h"
-SCALAR_TYPE_VERIFIER = SCALAR_TYPE_STEM + "_verifier.h"
+COMMON_BUILDER = COMMON_STEM + "_builder.h"
+COMMON_READER = COMMON_STEM + "_reader.h"
+COMMON_VERIFIER = COMMON_STEM + "_verifier.h"
 
 FLATBUFFERS_COMMON_STEM = "flatbuffers_common"
 FLATBUFFERS_COMMON_BUILDER = FLATBUFFERS_COMMON_STEM + "_builder.h"
@@ -52,14 +52,14 @@ def define_common_targets():
 
     generate_schema_header_flatcc(
         ETDUMP_GEN_RULE_NAME_FLATCC,
-        [ETDUMP_SCHEMA_FLATCC, SCALAR_TYPE],
+        [ETDUMP_SCHEMA_FLATCC, COMMON],
         [
             ETDUMP_SCHEMA_FLATCC_BUILDER,
             ETDUMP_SCHEMA_FLATCC_READER,
             ETDUMP_SCHEMA_FLATCC_VERIFIER,
-            SCALAR_TYPE_BUILDER,
-            SCALAR_TYPE_READER,
-            SCALAR_TYPE_VERIFIER,
+            COMMON_BUILDER,
+            COMMON_READER,
+            COMMON_VERIFIER,
             FLATBUFFERS_COMMON_BUILDER,
             FLATBUFFERS_COMMON_READER,
         ],
@@ -78,9 +78,9 @@ def define_common_targets():
             ETDUMP_SCHEMA_FLATCC_BUILDER: ":{}[{}]".format(ETDUMP_GEN_RULE_NAME_FLATCC, ETDUMP_SCHEMA_FLATCC_BUILDER),
             ETDUMP_SCHEMA_FLATCC_READER: ":{}[{}]".format(ETDUMP_GEN_RULE_NAME_FLATCC, ETDUMP_SCHEMA_FLATCC_READER),
             ETDUMP_SCHEMA_FLATCC_VERIFIER: ":{}[{}]".format(ETDUMP_GEN_RULE_NAME_FLATCC, ETDUMP_SCHEMA_FLATCC_VERIFIER),
-            SCALAR_TYPE_BUILDER: ":{}[{}]".format(ETDUMP_GEN_RULE_NAME_FLATCC, SCALAR_TYPE_BUILDER),
-            SCALAR_TYPE_READER: ":{}[{}]".format(ETDUMP_GEN_RULE_NAME_FLATCC, SCALAR_TYPE_READER),
-            SCALAR_TYPE_VERIFIER: ":{}[{}]".format(ETDUMP_GEN_RULE_NAME_FLATCC, SCALAR_TYPE_VERIFIER),
+            COMMON_BUILDER: ":{}[{}]".format(ETDUMP_GEN_RULE_NAME_FLATCC, COMMON_BUILDER),
+            COMMON_READER: ":{}[{}]".format(ETDUMP_GEN_RULE_NAME_FLATCC, COMMON_READER),
+            COMMON_VERIFIER: ":{}[{}]".format(ETDUMP_GEN_RULE_NAME_FLATCC, COMMON_VERIFIER),
             FLATBUFFERS_COMMON_BUILDER: ":{}[{}]".format(ETDUMP_GEN_RULE_NAME_FLATCC, FLATBUFFERS_COMMON_BUILDER),
             FLATBUFFERS_COMMON_READER: ":{}[{}]".format(ETDUMP_GEN_RULE_NAME_FLATCC, FLATBUFFERS_COMMON_READER),
         },
