+++ exir/_serialize/TARGETS
@@ -36,7 +36,7 @@ runtime.python_library(
     ],
     resources = {
         "//executorch/schema:program.fbs": "program.fbs",
-        "//executorch/schema:scalar_type.fbs": "scalar_type.fbs",
+        "//executorch/schema:common.fbs": "common.fbs",
         "fbsource//third-party/flatbuffers:flatc-host": "flatbuffers-flatc",
     },
     # Currently serialization API should only be used in some dedicated targets,
