+++ exir/_serialize/test/test_flatbuffer.py
@@ -5,6 +5,8 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
+# pyre-unsafe
+
 import os
 import re
 import shutil
@@ -85,7 +87,7 @@ SCHEMA_FILES: Dict[str, bytes] = {
             b"}",
         ]
     ),
-    "scalar_type.fbs": b"\n".join(
+    "common.fbs": b"\n".join(
         [
             b"table ScalarType {",
             # Spaces around the colon.
@@ -156,7 +158,7 @@ class TestPrepareSchema(unittest.TestCase):
                 ),
             )
             self.assertEqual(
-                read_file(out_dir, "scalar_type.fbs"),
+                read_file(out_dir, "common.fbs"),
                 b"\n".join(
                     [
                         b"table ScalarType {",
@@ -192,7 +194,7 @@ class TestPrepareSchema(unittest.TestCase):
                 ),
             )
             self.assertEqual(
-                read_file(out_dir, "scalar_type.fbs"),
+                read_file(out_dir, "common.fbs"),
                 b"\n".join(
                     [
                         b"table ScalarType {",
@@ -232,7 +234,7 @@ class TestPrepareSchema(unittest.TestCase):
                 ),
             )
             self.assertEqual(
-                read_file(out_dir, "scalar_type.fbs"),
+                read_file(out_dir, "common.fbs"),
                 b"\n".join(
                     [
                         b"table ScalarType {",
