+++ schema/program.fbs
@@ -4,7 +4,7 @@
 // See README.md before modifying this file.
 //
 
-include "scalar_type.fbs";
+include "common.fbs";
 
 namespace executorch_flatbuffer;
